refactor: fe - object 관련 타입 분리 #111
- workspace-object.d.ts에 object 관련 타입 선언
