+++ frontend/src/utils/object-from-server.ts
@@ -1,4 +1,3 @@
-import { CanvasObject } from '@pages/workspace/whiteboard-canvas/index.types';
 import { fabric } from 'fabric';
 
 export const createObjectFromServer = (canvas: fabric.Canvas, newObject: CanvasObject) => {
